Rasterize following crop and reproject

A simple wrapper around terra::rasterize().

fasterize2(emptyRaster, polygonToFasterize, field)

Arguments

emptyRaster

An empty RasterLayer orSpatRaster to use as a template.

polygonToFasterize

an sf or SpatVector object, which will be cropped first if extent(emptyRaster) < extent(polygonToFasterize).

field

character or numeric. If field is a character, it should a variable name in x. If field is numeric it typically is a single number or a vector of length nrow(x). The values are recycled to nrow(x)

Value

an object of the same class as emptyRaster

Examples

## using sf + raster
f1 <- system.file("external/lux.shp", package = "raster")
v1 <- sf::st_read(f1)
#> Reading layer `lux' from data source 
#>   `/home/runner/work/_temp/Library/raster/external/lux.shp' using driver `ESRI Shapefile'
#> Simple feature collection with 12 features and 5 fields
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 5.74414 ymin: 49.44781 xmax: 6.528252 ymax: 50.18162
#> Geodetic CRS:  WGS 84
r1 <- raster::raster(v1, ncols = 75, nrows = 100)
raster::crs(r1) <- "epsg:4326"
z1 <- fasterize2(r1, v1, "NAME_2")
#> projecting...
#> done! took:  0.00281 secs

## using terra
f2 <- system.file("ex/lux.shp", package = "terra")
v2 <- terra::vect(f2)
r2 <- terra::rast(v2, ncols = 75, nrows = 100)
z2 <- fasterize2(r2, v2, "NAME_2")
#> projecting...
#> done! took:  0.00197 secs

terra::compareGeom(terra::rast(z1), z2)
#> [1] TRUE