The map class

Contains a common system for organzing vector and raster layers, principally for use with leaflet and shiny.

Slots

metadata

data.table with columns describing metadata of map objects in maps slot.

.xData

Named environment of map-type objects (e.g., sf, Raster*, Spatial*. Each entry may also be simply an environment, which indicates where to find the object, i.e., via get(layerName, envir = environment)

CRS

The common crs of all layers

paths

File paths. A named list of paths. The default is a list of length 2, dataPath and tilePath

analyses

A data.table or data.frame of the types of analyses to perform.

analysesData

A data.table or data.frame of the results of the analyses.