Utility functions for grouping analyses in a map object

Utility functions for grouping analyses in a map object

rbindlistAG(map, functionName, analysisGroups)

Arguments

map

Optional map object. If not provided, then one will be created. If provided, then the present obj or options passed to reproducible::prepInputs() (e.g., url), will be appended to this map.

functionName

TODO: description needed

analysisGroups

A character (length 1 currently), indicating which analysis group (e.g., "analysisGroup1") should be used to rbindlist. Can also specify "all" which will rbindlist all outputs.

Value

A list of data.tables.