Extract rasters in the map object

This will extract all objects in or pointed to within the map.

rasters(map)

# S3 method for map
rasters(map)

sp(map)

# S3 method for map
sp(map)

sf(map)

# S3 method for map
sf(map)

spatialPolygons(map)

spatialPoints(map)

maps(map, class = NULL, layerName = NULL)

Arguments

map

A map class obj

class

If supplied, this will be the class of objects returned. Default is NULL which is "all", meaning all objects in the map object.

layerName

TODO: description needed

Value

A list of maps (i.e., sp, raster, or sf objects) of class class

See also