Extract objects of specific classes from a map object

rasters() extracts RasterLayer objects;

sf() extracts sf objects;

sp() extracts Spatial objects;

spatialPolygons() extracts SpatialPolygons objects;

spatialPoints() extracts SpatialPoints objects;

spatRasters() extracts SpatRaster objects;

spatVectors() extracts SpatVector objects;

This will extract all objects in or pointed to within the map.

rasters(map, ...)

# S3 method for class 'map'
rasters(map, ...)

sp(map, ...)

# S3 method for class 'map'
sp(map, ...)

sf(map, ...)

# S3 method for class 'map'
sf(map, ...)

spatialPolygons(map, ...)

# S3 method for class 'map'
spatialPolygons(map, ...)

spatialPoints(map, ...)

# S3 method for class 'map'
spatialPoints(map, ...)

spatRasters(map, ...)

# S3 method for class 'map'
spatRasters(map, ...)

spatVectors(map, ...)

# S3 method for class 'map'
spatVectors(map, ...)

maps(map, class = NULL, layerName = NULL)

Arguments

map

A map object

...

Additional arguments passed to other methods (not used)

class

If supplied, this will be the class of objects returned. Default is NULL which is "all", meaning all objects in the map object.

layerName

character string giving the name(s) of layer(s) to extract.

Value

A list of geospatial objects of class class

See also